+++ frontend/src/cards/ui/AltTableView.tsx
@@ -10,10 +10,7 @@ import {
 import WarningRoundedIcon from '@mui/icons-material/WarningRounded'
 import { useRef } from 'react'
 import AnimateHeight from 'react-animate-height'
-import {
-  formatFieldValue,
-  type MetricConfig,
-} from '../../data/config/MetricConfig'
+import type { MetricConfig } from '../../data/config/MetricConfig'
 import type { DemographicType } from '../../data/query/Breakdowns'
 import {
   type DemographicGroup,
@@ -29,7 +26,10 @@ import {
   ALT_TABLE_VIEW_2_PARAM_KEY,
 } from '../../utils/urlutils'
 import HetExpandableBoxButton from '../../styles/HetComponents/HetExpandableBoxButton'
-import { isPctType } from '../../data/config/MetricConfigUtils'
+import {
+  formatFieldValue,
+  isPctType,
+} from '../../data/config/MetricConfigUtils'
 
 interface AltTableViewProps {
   expanded: boolean
