+++ frontend/src/cards/ui/ExtremeList.tsx
@@ -1,10 +1,8 @@
 import { Grid } from '@mui/material'
-import {
-  type MetricConfig,
-  formatFieldValue,
-} from '../../data/config/MetricConfig'
+import type { MetricConfig } from '../../data/config/MetricConfig'
 import type { Row } from '../../data/utils/DatasetTypes'
 import HetUnitLabel from '../../styles/HetComponents/HetUnitLabel'
+import { formatFieldValue } from '../../data/config/MetricConfigUtils'
 
 interface ExtremeListProps {
   whichExtreme: 'Highest' | 'Lowest'
