+++ frontend/src/cards/ui/ExtremesListBox.tsx
@@ -1,9 +1,8 @@
 import AnimateHeight from 'react-animate-height'
 import { Grid } from '@mui/material'
-import {
-  type MetricConfig,
-  type DataTypeConfig,
-  formatFieldValue,
+import type {
+  MetricConfig,
+  DataTypeConfig,
 } from '../../data/config/MetricConfig'
 import type { Row } from '../../data/utils/DatasetTypes'
 import { WHAT_DATA_ARE_MISSING_ID } from '../../utils/internalRoutes'
@@ -15,6 +14,7 @@ import ExtremeList from './ExtremeList'
 import HetUnitLabel from '../../styles/HetComponents/HetUnitLabel'
 import HetTerm from '../../styles/HetComponents/HetTerm'
 import HetExpandableBoxButton from '../../styles/HetComponents/HetExpandableBoxButton'
+import { formatFieldValue } from '../../data/config/MetricConfigUtils'
 
 interface ExtremesListBoxProps {
   // MetricConfig for data
