+++ frontend/src/charts/utils.ts
@@ -1,8 +1,5 @@
-import {
-  type DataTypeConfig,
-  formatFieldValue,
-  type MetricConfig,
-} from '../data/config/MetricConfig'
+import type { DataTypeConfig, MetricConfig } from '../data/config/MetricConfig'
+import { formatFieldValue } from '../data/config/MetricConfigUtils'
 import {
   DEMOGRAPHIC_DISPLAY_TYPES_LOWER_CASE,
   type DemographicType,
