+++ frontend/src/data/config/MetricConfigUtils.test.ts
@@ -6,7 +6,11 @@ import type {
   MetricId,
   MetricType,
 } from './MetricConfig'
-import { isPctType, metricConfigFromDtConfig } from './MetricConfigUtils'
+import {
+  formatFieldValue,
+  isPctType,
+  metricConfigFromDtConfig,
+} from './MetricConfigUtils'
 
 describe('metricConfigFromDtConfig', () => {
   const fakeDataTypeConfig: DataTypeConfig = {
@@ -69,4 +73,13 @@ describe('Test Metric Config Functions', () => {
     expect(isPctType('per100k')).toBe(false)
     expect(isPctType('something broken' as MetricType)).toBe(false)
   })
+})
+
+test('Test Formatting of Field Values', () => {
+  expect(formatFieldValue('pct_relative_inequity', 33)).toBe('33%')
+  expect(formatFieldValue('pct_rate', 33, true)).toBe('33')
+  expect(formatFieldValue('pct_share', 3, false)).toBe('3.0%')
+  expect(formatFieldValue('per100k', 30_000, false)).toBe('30,000')
+  expect(formatFieldValue('per100k', 0, false)).toBe('< 0.1')
+  expect(formatFieldValue('per100k', 3.33, false)).toBe('3.3')
 })
